+mensajes entre usuarios al reservar/vender inmueble
