+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Android-Studio-Projects/InmuebleLibre/app/src/main/res/layout/activity_agregar_inmueble.xml" value="0.36666666666666664" />
         <entry key="..\:/Android-Studio-Projects/InmuebleLibre/app/src/main/res/layout/activity_configuracion_cuenta.xml" value="0.36666666666666664" />
+        <entry key="..\:/Android-Studio-Projects/InmuebleLibre/app/src/main/res/layout/activity_enviar_notificaciones.xml" value="0.18697916666666667" />
         <entry key="..\:/Android-Studio-Projects/InmuebleLibre/app/src/main/res/layout/activity_inicio_cliente.xml" value="0.35260416666666666" />
         <entry key="..\:/Android-Studio-Projects/InmuebleLibre/app/src/main/res/layout/activity_inicio_propietario.xml" value="0.22" />
         <entry key="..\:/Android-Studio-Projects/InmuebleLibre/app/src/main/res/layout/activity_lista_comprados_cliente.xml" value="0.19427083333333334" />
