+++ app/src/main/AndroidManifest.xml
@@ -15,46 +15,46 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.InmuebleLibre">
         <activity
-            android:name=".UI.global.VerDetallesInmuebleActivity"
+            android:name=".ui.global.VerDetallesInmuebleActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.cliente.ListaCompradosClienteActivity"
+            android:name=".ui.cliente.ListaCompradosClienteActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.global.ConfiguracionCuentaActivity"
+            android:name=".ui.global.ConfiguracionCuentaActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.propietario.ListaVendidosPropietarioActivity"
+            android:name=".ui.propietario.ListaVendidosPropietarioActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.propietario.ListaReservasPropietarioActivity"
+            android:name=".ui.propietario.ListaReservasPropietarioActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.cliente.ListaInmueblesClienteActivity"
+            android:name=".ui.cliente.ListaInmueblesClienteActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.cliente.InicioClienteActivity"
+            android:name=".ui.cliente.InicioClienteActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.propietario.InicioPropietarioActivity"
+            android:name=".ui.propietario.InicioPropietarioActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.global.LoginActivity"
+            android:name=".ui.global.LoginActivity"
             android:exported="true"
-            android:parentActivityName=".UI.global.MainActivity" />
+            android:parentActivityName=".ui.global.MainActivity" />
         <activity
-            android:name=".UI.global.RegistrarActivity"
+            android:name=".ui.global.RegistrarActivity"
             android:exported="true"
-            android:parentActivityName=".UI.global.MainActivity" />
+            android:parentActivityName=".ui.global.MainActivity" />
         <activity
-            android:name=".UI.propietario.AgregarInmuebleActivity"
+            android:name=".ui.propietario.AgregarInmuebleActivity"
             android:exported="true"
-            android:parentActivityName=".UI.propietario.InicioPropietarioActivity" />
+            android:parentActivityName=".ui.propietario.InicioPropietarioActivity" />
         <activity
-            android:name=".UI.propietario.MapsActivity"
+            android:name=".ui.propietario.MapsActivity"
             android:exported="true" />
         <activity
-            android:name=".UI.global.MainActivity"
+            android:name=".ui.global.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -63,6 +63,17 @@
             </intent-filter>
         </activity>
 
+        <service
+            android:name=".ServicioFirebaseMensajes"
+            android:exported="false">
+            <intent-filter>
+                <action android:name="com.google.firebase.MESSAGING_EVENT" />
+            </intent-filter>
+        </service>
+
+        <meta-data
+            android:name="com.google.firebase.messaging.default_notification_channel_id"
+            android:value="@string/default_notification_channel_id" />
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="${MAPS_API_KEY}" />
