+++ app/src/main/java/com/mmw/inmueblelibre/apidao/BuilderAPI.java
@@ -10,6 +10,7 @@ import retrofit2.converter.gson.GsonConverterFactory;
 public class BuilderAPI {
     private static BuilderAPI _INSTANCIA;
     private CiudadesAPI ciudadesAPI;
+    private MensajesFirebaseAPI mensajesFirebaseAPI;
 
     private void iniciarRetrofit(){
         Gson gson = new GsonBuilder().setLenient().create();
@@ -20,7 +21,16 @@ public class BuilderAPI {
                 .client(cliente)
                 .addConverterFactory(GsonConverterFactory.create(gson))
                 .build();
+
+        Retrofit retrofitMensajes = new Retrofit.Builder()
+                .baseUrl("https://fcm.googleapis.com/fcm/")
+                .client(cliente)
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .build();
+
+
         ciudadesAPI = retrofit.create(CiudadesAPI.class);
+        mensajesFirebaseAPI = retrofitMensajes.create(MensajesFirebaseAPI.class);
     }
 
     public static BuilderAPI getInstancia(){
@@ -35,5 +45,12 @@ public class BuilderAPI {
             this.iniciarRetrofit();
         }
         return ciudadesAPI;
+    }
+
+    public MensajesFirebaseAPI getMensajesFirebaseAPI(){
+        if (mensajesFirebaseAPI == null) {
+            this.iniciarRetrofit();
+        }
+        return mensajesFirebaseAPI;
     }
 }
