+++ app/src/main/java/com/mmw/inmueblelibre/ui/cliente/InicioClienteActivity.java
@@ -1,4 +1,4 @@
-package com.mmw.inmueblelibre.UI.cliente;
+package com.mmw.inmueblelibre.ui.cliente;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,6 +14,7 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
@@ -22,17 +23,19 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.mmw.inmueblelibre.UI.global.MainActivity;
+import com.mmw.inmueblelibre.ui.global.MainActivity;
 import com.mmw.inmueblelibre.R;
-import com.mmw.inmueblelibre.UI.global.ConfiguracionCuentaActivity;
-import com.mmw.inmueblelibre.UI.global.VerDetallesInmuebleActivity;
+import com.mmw.inmueblelibre.ui.global.ConfiguracionCuentaActivity;
+import com.mmw.inmueblelibre.ui.global.VerDetallesInmuebleActivity;
 import com.mmw.inmueblelibre.adapter.InmuebleAdapter;
 import com.mmw.inmueblelibre.model.InmuebleModel;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public class ListaCompradosClienteActivity extends AppCompatActivity {
+public class InicioClienteActivity extends AppCompatActivity {
 
     Toolbar toolbar;
 
@@ -42,26 +45,26 @@ public class ListaCompradosClienteActivity extends AppCompatActivity {
     private DrawerLayout drawerLayout;
     private NavigationView menuDrawer;
 
-    RecyclerView.LayoutManager inmueblesLayoutManager;
-    RecyclerView listaInmueblesRV;
-    InmuebleAdapter adaptadorListaInmuebles;
+    RecyclerView.LayoutManager reservasLayoutManager;
+    RecyclerView listaReservasRV;
+    InmuebleAdapter adaptadorListaReservas;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_lista_comprados_cliente);
+        setContentView(R.layout.activity_inicio_cliente);
 
-        toolbar = findViewById(R.id.LCC_toolbar);
+        toolbar = findViewById(R.id.INICIOCLI_toolbar);
         setSupportActionBar(toolbar);
 
-        drawerLayout = (DrawerLayout) findViewById(R.id.LCC_drawer_layout);
-        menuDrawer = (NavigationView) findViewById(R.id.LCC_menu_drawer);
+        drawerLayout = (DrawerLayout) findViewById(R.id.INICIOCLI_drawer_layout);
+        menuDrawer = (NavigationView) findViewById(R.id.INICIOCLI_menu_drawer);
 
-        listaInmueblesRV = findViewById(R.id.LCC_listaInmuebles);
-        inmueblesLayoutManager = new LinearLayoutManager(this);
-        listaInmueblesRV.setLayoutManager(inmueblesLayoutManager);
-        adaptadorListaInmuebles = new InmuebleAdapter(new ArrayList<>());
-        listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
+        listaReservasRV = findViewById(R.id.INICIOCLI_listaReservas);
+        reservasLayoutManager = new LinearLayoutManager(this);
+        listaReservasRV.setLayoutManager(reservasLayoutManager);
+        adaptadorListaReservas = new InmuebleAdapter(new ArrayList<>());
+        listaReservasRV.setAdapter(adaptadorListaReservas);
 
         firebaseAuth = FirebaseAuth.getInstance();
         databaseFirebase = FirebaseDatabase.getInstance().getReference();
@@ -76,41 +79,56 @@ public class ListaCompradosClienteActivity extends AppCompatActivity {
 
             switch (menuItem.getItemId()) {
                 case R.id.MENUCLI_menu_principal_opc:
-                    startActivity(new Intent(ListaCompradosClienteActivity.this, InicioClienteActivity.class));
                     break;
 
                 case R.id.MENUCLI_listar_inmuebles_opc:
-                    startActivity(new Intent(ListaCompradosClienteActivity.this, ListaInmueblesClienteActivity.class));
+                    startActivity(new Intent(InicioClienteActivity.this, ListaInmueblesClienteActivity.class));
                     break;
 
                 case R.id.MENUCLI_listar_comprados_opc:
+                    startActivity(new Intent(InicioClienteActivity.this, ListaCompradosClienteActivity.class));
                     break;
 
                 case R.id.MENUCLI_mi_cuenta_opc:
-                    startActivity(new Intent(ListaCompradosClienteActivity.this, ConfiguracionCuentaActivity.class));
+                    startActivity(new Intent(InicioClienteActivity.this, ConfiguracionCuentaActivity.class));
                     break;
 
                 case R.id.MENUCLI_cerrar_sesion_opc:
-                    firebaseAuth.signOut();
-                    startActivity(new Intent(ListaCompradosClienteActivity.this, MainActivity.class));
-                    finish();
+                    cerrarSesion();
                     break;
             }
 
             return false;
         });
 
-        adaptadorListaInmuebles.setOnItemClickListener(new InmuebleAdapter.OnItemClickListener() {
+        adaptadorListaReservas.setOnItemClickListener(new InmuebleAdapter.OnItemClickListener() {
             @Override
             public void onItemClick(View itemView, int position) {
                 String tipoUsuario = "CLIENTE";
-                String idInmueble = adaptadorListaInmuebles.getInmuebleId(position);
+                String idInmueble = adaptadorListaReservas.getInmuebleId(position);
 
-                Intent intent = new Intent(ListaCompradosClienteActivity.this, VerDetallesInmuebleActivity.class);
+                Intent intent = new Intent(InicioClienteActivity.this, VerDetallesInmuebleActivity.class);
                 intent.putExtra("tipo_usuario", tipoUsuario);
                 intent.putExtra("id_inmueble", idInmueble);
-                intent.putExtra("estado_inmueble", "VENDIDO");
+                intent.putExtra("estado_inmueble", "RESERVADO");
                 startActivity(intent);
+            }
+        });
+    }
+
+    private void cerrarSesion(){
+        String id = firebaseAuth.getCurrentUser().getUid();
+
+        Map<String, Object> mapaValores = new HashMap<>();
+        mapaValores.put("token_fcm", "");
+
+        databaseFirebase.child("Usuarios").child(id).updateChildren(mapaValores).addOnCompleteListener(taskDB -> {
+            if (taskDB.isSuccessful()){
+                firebaseAuth.signOut();
+                startActivity(new Intent(InicioClienteActivity.this, MainActivity.class));
+                finish();
+            } else {
+                Toast.makeText(getApplicationContext(), "No se pudo actualizar el token", Toast.LENGTH_SHORT).show();
             }
         });
     }
@@ -160,27 +178,28 @@ public class ListaCompradosClienteActivity extends AppCompatActivity {
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
                 if (snapshot.exists()){
-                    List<InmuebleModel> listaInmuebles = new ArrayList<>();
+                    List<InmuebleModel> listaReservas = new ArrayList<>();
 
                     for (DataSnapshot snap : snapshot.getChildren()){
-                        String id = snap.getKey();
+                        String id = snap.getKey().toString();
+
                         String idActual = firebaseAuth.getCurrentUser().getUid();
 
+                        String idClienteInmueble = snap.child("id_cliente").getValue().toString();
                         String estado = snap.child("estado").getValue().toString();
                         String precio = snap.child("precio").getValue().toString();
                         String direccion = (snap.child("direccion").getValue().toString().split(": "))[1];
-                        String idCliente = snap.child("id_cliente").getValue().toString();
 
-                        if (!idActual.equals(idCliente)) continue;
-                        if (!estado.equals("VENDIDO")) continue;
+                        if (!idActual.equals(idClienteInmueble)) continue;
+                        if (!estado.equals("RESERVADO")) continue;
 
                         InmuebleModel inmTemp = new InmuebleModel(id, direccion, precio);
 
-                        listaInmuebles.add(inmTemp);
+                        listaReservas.add(inmTemp);
                     }
 
-                    adaptadorListaInmuebles = new InmuebleAdapter(listaInmuebles);
-                    listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
+                    adaptadorListaReservas = new InmuebleAdapter(listaReservas);
+                    listaReservasRV.setAdapter(adaptadorListaReservas);
                 }
             }
