+++ app/src/main/java/com/mmw/inmueblelibre/ui/cliente/ListaInmueblesClienteActivity.java
@@ -1,4 +1,4 @@
-package com.mmw.inmueblelibre.UI.propietario;
+package com.mmw.inmueblelibre.ui.cliente;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -16,7 +16,6 @@ import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
@@ -24,17 +23,19 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.mmw.inmueblelibre.UI.global.MainActivity;
+import com.mmw.inmueblelibre.ui.global.MainActivity;
 import com.mmw.inmueblelibre.R;
-import com.mmw.inmueblelibre.UI.global.ConfiguracionCuentaActivity;
-import com.mmw.inmueblelibre.UI.global.VerDetallesInmuebleActivity;
+import com.mmw.inmueblelibre.ui.global.ConfiguracionCuentaActivity;
+import com.mmw.inmueblelibre.ui.global.VerDetallesInmuebleActivity;
 import com.mmw.inmueblelibre.adapter.InmuebleAdapter;
 import com.mmw.inmueblelibre.model.InmuebleModel;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public class InicioPropietarioActivity extends AppCompatActivity implements View.OnClickListener {
+public class ListaInmueblesClienteActivity extends AppCompatActivity {
 
     Toolbar toolbar;
 
@@ -48,78 +49,86 @@ public class InicioPropietarioActivity extends AppCompatActivity implements View
     RecyclerView listaInmueblesRV;
     InmuebleAdapter adaptadorListaInmuebles;
 
-    private FloatingActionButton agregarInmuebleBtn;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_inicio_propietario);
+        setContentView(R.layout.activity_lista_inmuebles_cliente);
 
-        toolbar = findViewById(R.id.INICIOPROP_toolbar);
+        toolbar = findViewById(R.id.LIC_toolbar);
         setSupportActionBar(toolbar);
 
-        firebaseAuth = FirebaseAuth.getInstance();
-        databaseFirebase = FirebaseDatabase.getInstance().getReference();
+        drawerLayout = (DrawerLayout) findViewById(R.id.LIC_drawer_layout);
+        menuDrawer = (NavigationView) findViewById(R.id.LIC_menu_drawer);
 
-        drawerLayout = (DrawerLayout) findViewById(R.id.INICIOPROP_drawer_layout);
-        menuDrawer = (NavigationView) findViewById(R.id.INICIOPROP_menu_drawer);
-
-        agregarInmuebleBtn = findViewById(R.id.INICIOPROP_agregar_inmueble_BTN);
-
-        listaInmueblesRV = findViewById(R.id.INICIOPROP_listaInmuebles);
+        listaInmueblesRV = findViewById(R.id.LIC_listaInmuebles);
         inmueblesLayoutManager = new LinearLayoutManager(this);
         listaInmueblesRV.setLayoutManager(inmueblesLayoutManager);
         adaptadorListaInmuebles = new InmuebleAdapter(new ArrayList<>());
         listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
+
+        firebaseAuth = FirebaseAuth.getInstance();
+        databaseFirebase = FirebaseDatabase.getInstance().getReference();
+
         obtenerLista();
 
+        obtenerInfoUsuario();
 
-        agregarInmuebleBtn.setOnClickListener(this);
-
-        //TODO IMPLEMENTAR ACCIONES DEL MENU DRAWER
         menuDrawer.setNavigationItemSelectedListener(menuItem -> {
 
             drawerLayout.closeDrawer(menuDrawer);
 
-            switch (menuItem.getItemId()){
-                case R.id.MENUPROP_menu_principal_opc:
+            switch (menuItem.getItemId()) {
+                case R.id.MENUCLI_menu_principal_opc:
+                    startActivity(new Intent(ListaInmueblesClienteActivity.this, InicioClienteActivity.class));
                     break;
 
-                case R.id.MENUPROP_listar_reservas_opc:
-                    startActivity(new Intent(InicioPropietarioActivity.this, ListaReservasPropietarioActivity.class));
+                case R.id.MENUCLI_listar_inmuebles_opc:
                     break;
 
-                case R.id.MENUPROP_listar_vendidos_opc:
-                    startActivity(new Intent(InicioPropietarioActivity.this, ListaVendidosPropietarioActivity.class));
+                case R.id.MENUCLI_listar_comprados_opc:
+                    startActivity(new Intent(ListaInmueblesClienteActivity.this, ListaCompradosClienteActivity.class));
                     break;
 
-                case R.id.MENUPROP_mi_cuenta_opc:
-                    startActivity(new Intent(InicioPropietarioActivity.this, ConfiguracionCuentaActivity.class));
+                case R.id.MENUCLI_mi_cuenta_opc:
+                    startActivity(new Intent(ListaInmueblesClienteActivity.this, ConfiguracionCuentaActivity.class));
                     break;
 
-                case R.id.MENUPROP_cerrar_sesion_opc:
-                    firebaseAuth.signOut();
-                    startActivity(new Intent(InicioPropietarioActivity.this, MainActivity.class));
-                    finish();
+                case R.id.MENUCLI_cerrar_sesion_opc:
+                    cerrarSesion();
                     break;
             }
 
             return false;
         });
 
-        obtenerInfoUsuario();
-
         adaptadorListaInmuebles.setOnItemClickListener(new InmuebleAdapter.OnItemClickListener() {
             @Override
             public void onItemClick(View itemView, int position) {
-                String tipoUsuario = "PROPIETARIO";
+                String tipoUsuario = "CLIENTE";
                 String idInmueble = adaptadorListaInmuebles.getInmuebleId(position);
 
-                Intent intent = new Intent(InicioPropietarioActivity.this, VerDetallesInmuebleActivity.class);
+                Intent intent = new Intent(ListaInmueblesClienteActivity.this, VerDetallesInmuebleActivity.class);
                 intent.putExtra("tipo_usuario", tipoUsuario);
                 intent.putExtra("id_inmueble", idInmueble);
                 intent.putExtra("estado_inmueble", "CREADO");
                 startActivity(intent);
+            }
+        });
+    }
+
+    private void cerrarSesion(){
+        String id = firebaseAuth.getCurrentUser().getUid();
+
+        Map<String, Object> mapaValores = new HashMap<>();
+        mapaValores.put("token_fcm", "");
+
+        databaseFirebase.child("Usuarios").child(id).updateChildren(mapaValores).addOnCompleteListener(taskDB -> {
+            if (taskDB.isSuccessful()){
+                firebaseAuth.signOut();
+                startActivity(new Intent(ListaInmueblesClienteActivity.this, MainActivity.class));
+                finish();
+            } else {
+                Toast.makeText(getApplicationContext(), "No se pudo actualizar el token", Toast.LENGTH_SHORT).show();
             }
         });
     }
@@ -141,49 +150,6 @@ public class InicioPropietarioActivity extends AppCompatActivity implements View
         return super.onOptionsItemSelected(item);
     }
 
-    @Override
-    public void onClick(View view) {
-        switch (view.getId()){
-            case R.id.INICIOPROP_agregar_inmueble_BTN:
-                startActivity(new Intent(InicioPropietarioActivity.this, AgregarInmuebleActivity.class));
-                break;
-        }
-    }
-
-    private void obtenerLista(){
-        databaseFirebase.child("Inmuebles").addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()){
-                    List<InmuebleModel> listaInmuebles = new ArrayList<>();
-
-                    for (DataSnapshot snap : snapshot.getChildren()){
-                        String id = snap.getKey().toString();
-                        String precio = snap.child("precio").getValue().toString();
-                        String direccion = (snap.child("direccion").getValue().toString().split(": "))[1];
-                        String idProp = snap.child("id_propietario").getValue().toString();
-                        String estado = snap.child("estado").getValue().toString();
-
-                        if (!idProp.equals(firebaseAuth.getCurrentUser().getUid())) continue;
-                        if (!estado.equals("CREADO")) continue;
-
-                        InmuebleModel inmTemp = new InmuebleModel(id, direccion, precio);
-
-                        listaInmuebles.add(inmTemp);
-
-                    }
-                    adaptadorListaInmuebles = new InmuebleAdapter(listaInmuebles);
-                    listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-
     private void obtenerInfoUsuario(){
         String id = firebaseAuth.getCurrentUser().getUid();
 
@@ -195,8 +161,41 @@ public class InicioPropietarioActivity extends AppCompatActivity implements View
                     String correo = snapshot.child("email").getValue().toString();
 
                     View drawerHeader = menuDrawer.getHeaderView(0);
-                    ((TextView) drawerHeader.findViewById(R.id.DRAWERPROP_nombre_TV)).setText(nombre);
-                    ((TextView) drawerHeader.findViewById(R.id.DRAWERPROP_correo_TV)).setText(correo);
+                    ((TextView) drawerHeader.findViewById(R.id.DRAWERCLI_nombre_TV)).setText(nombre);
+                    ((TextView) drawerHeader.findViewById(R.id.DRAWERCLI_correo_TV)).setText(correo);
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+
+            }
+        });
+    }
+
+    private void obtenerLista(){
+        databaseFirebase.child("Inmuebles").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                if (snapshot.exists()){
+                    List<InmuebleModel> listaInmuebles = new ArrayList<>();
+
+                    for (DataSnapshot snap : snapshot.getChildren()){
+                        String id = snap.getKey();
+
+                        String estado = snap.child("estado").getValue().toString();
+                        String precio = snap.child("precio").getValue().toString();
+                        String direccion = (snap.child("direccion").getValue().toString().split(": "))[1];
+
+                        if (!estado.equals("CREADO")) continue;
+
+                        InmuebleModel inmTemp = new InmuebleModel(id, direccion, precio);
+
+                        listaInmuebles.add(inmTemp);
+                    }
+
+                    adaptadorListaInmuebles = new InmuebleAdapter(listaInmuebles);
+                    listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
                 }
             }
