+++ app/src/main/java/com/mmw/inmueblelibre/ui/global/ConfiguracionCuentaActivity.java
@@ -1,4 +1,4 @@
-package com.mmw.inmueblelibre.UI.global;
+package com.mmw.inmueblelibre.ui.global;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
@@ -20,8 +20,8 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.mmw.inmueblelibre.UI.cliente.InicioClienteActivity;
-import com.mmw.inmueblelibre.UI.propietario.InicioPropietarioActivity;
+import com.mmw.inmueblelibre.ui.cliente.InicioClienteActivity;
+import com.mmw.inmueblelibre.ui.propietario.InicioPropietarioActivity;
 import com.mmw.inmueblelibre.R;
 
 import java.util.HashMap;
