+++ app/src/main/java/com/mmw/inmueblelibre/ui/propietario/InicioPropietarioActivity.java
@@ -1,4 +1,4 @@
-package com.mmw.inmueblelibre.UI.propietario;
+package com.mmw.inmueblelibre.ui.propietario;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,7 +14,9 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
@@ -22,17 +24,20 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.mmw.inmueblelibre.UI.global.MainActivity;
+import com.mmw.inmueblelibre.ui.cliente.InicioClienteActivity;
+import com.mmw.inmueblelibre.ui.global.MainActivity;
 import com.mmw.inmueblelibre.R;
-import com.mmw.inmueblelibre.UI.global.ConfiguracionCuentaActivity;
-import com.mmw.inmueblelibre.UI.global.VerDetallesInmuebleActivity;
+import com.mmw.inmueblelibre.ui.global.ConfiguracionCuentaActivity;
+import com.mmw.inmueblelibre.ui.global.VerDetallesInmuebleActivity;
 import com.mmw.inmueblelibre.adapter.InmuebleAdapter;
 import com.mmw.inmueblelibre.model.InmuebleModel;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public class ListaVendidosPropietarioActivity extends AppCompatActivity {
+public class InicioPropietarioActivity extends AppCompatActivity implements View.OnClickListener {
 
     Toolbar toolbar;
 
@@ -42,30 +47,37 @@ public class ListaVendidosPropietarioActivity extends AppCompatActivity {
     private DrawerLayout drawerLayout;
     private NavigationView menuDrawer;
 
-    RecyclerView.LayoutManager vendidosLayoutManager;
-    RecyclerView listaVendidosRV;
-    InmuebleAdapter adaptadorListaVendidos;
+    RecyclerView.LayoutManager inmueblesLayoutManager;
+    RecyclerView listaInmueblesRV;
+    InmuebleAdapter adaptadorListaInmuebles;
+
+    private FloatingActionButton agregarInmuebleBtn;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_lista_vendidos_propietario);
+        setContentView(R.layout.activity_inicio_propietario);
 
-        toolbar = findViewById(R.id.LVP_toolbar);
+        toolbar = findViewById(R.id.INICIOPROP_toolbar);
         setSupportActionBar(toolbar);
 
         firebaseAuth = FirebaseAuth.getInstance();
         databaseFirebase = FirebaseDatabase.getInstance().getReference();
 
-        drawerLayout = (DrawerLayout) findViewById(R.id.LVP_drawer_layout);
-        menuDrawer = (NavigationView) findViewById(R.id.LVP_menu_drawer);
+        drawerLayout = (DrawerLayout) findViewById(R.id.INICIOPROP_drawer_layout);
+        menuDrawer = (NavigationView) findViewById(R.id.INICIOPROP_menu_drawer);
 
-        listaVendidosRV = findViewById(R.id.LVP_listaInmuebles);
-        vendidosLayoutManager = new LinearLayoutManager(this);
-        listaVendidosRV.setLayoutManager(vendidosLayoutManager);
-        adaptadorListaVendidos = new InmuebleAdapter(new ArrayList<>());
-        listaVendidosRV.setAdapter(adaptadorListaVendidos);
+        agregarInmuebleBtn = findViewById(R.id.INICIOPROP_agregar_inmueble_BTN);
+
+        listaInmueblesRV = findViewById(R.id.INICIOPROP_listaInmuebles);
+        inmueblesLayoutManager = new LinearLayoutManager(this);
+        listaInmueblesRV.setLayoutManager(inmueblesLayoutManager);
+        adaptadorListaInmuebles = new InmuebleAdapter(new ArrayList<>());
+        listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
         obtenerLista();
+
+
+        agregarInmuebleBtn.setOnClickListener(this);
 
         //TODO IMPLEMENTAR ACCIONES DEL MENU DRAWER
         menuDrawer.setNavigationItemSelectedListener(menuItem -> {
@@ -74,24 +86,22 @@ public class ListaVendidosPropietarioActivity extends AppCompatActivity {
 
             switch (menuItem.getItemId()){
                 case R.id.MENUPROP_menu_principal_opc:
-                    startActivity(new Intent(ListaVendidosPropietarioActivity.this, InicioPropietarioActivity.class));
                     break;
 
                 case R.id.MENUPROP_listar_reservas_opc:
-                    startActivity(new Intent(ListaVendidosPropietarioActivity.this, ListaReservasPropietarioActivity.class));
+                    startActivity(new Intent(InicioPropietarioActivity.this, ListaReservasPropietarioActivity.class));
                     break;
 
                 case R.id.MENUPROP_listar_vendidos_opc:
+                    startActivity(new Intent(InicioPropietarioActivity.this, ListaVendidosPropietarioActivity.class));
                     break;
 
                 case R.id.MENUPROP_mi_cuenta_opc:
-                    startActivity(new Intent(ListaVendidosPropietarioActivity.this, ConfiguracionCuentaActivity.class));
+                    startActivity(new Intent(InicioPropietarioActivity.this, ConfiguracionCuentaActivity.class));
                     break;
 
                 case R.id.MENUPROP_cerrar_sesion_opc:
-                    firebaseAuth.signOut();
-                    startActivity(new Intent(ListaVendidosPropietarioActivity.this, MainActivity.class));
-                    finish();
+                    cerrarSesion();
                     break;
             }
 
@@ -100,19 +110,38 @@ public class ListaVendidosPropietarioActivity extends AppCompatActivity {
 
         obtenerInfoUsuario();
 
-        adaptadorListaVendidos.setOnItemClickListener(new InmuebleAdapter.OnItemClickListener() {
+        adaptadorListaInmuebles.setOnItemClickListener(new InmuebleAdapter.OnItemClickListener() {
             @Override
             public void onItemClick(View itemView, int position) {
                 String tipoUsuario = "PROPIETARIO";
-                String idInmueble = adaptadorListaVendidos.getInmuebleId(position);
+                String idInmueble = adaptadorListaInmuebles.getInmuebleId(position);
 
-                Intent intent = new Intent(ListaVendidosPropietarioActivity.this, VerDetallesInmuebleActivity.class);
+                Intent intent = new Intent(InicioPropietarioActivity.this, VerDetallesInmuebleActivity.class);
                 intent.putExtra("tipo_usuario", tipoUsuario);
                 intent.putExtra("id_inmueble", idInmueble);
-                intent.putExtra("estado_inmueble", "VENDIDO");
+                intent.putExtra("estado_inmueble", "CREADO");
                 startActivity(intent);
             }
         });
+    }
+
+    private void cerrarSesion(){
+        String id = firebaseAuth.getCurrentUser().getUid();
+
+        Map<String, Object> mapaValores = new HashMap<>();
+        mapaValores.put("token_fcm", "");
+
+        databaseFirebase.child("Usuarios").child(id).updateChildren(mapaValores).addOnCompleteListener(taskDB -> {
+            if (taskDB.isSuccessful()){
+                firebaseAuth.signOut();
+                startActivity(new Intent(InicioPropietarioActivity.this, MainActivity.class));
+                finish();
+            } else {
+                Toast.makeText(getApplicationContext(), "No se pudo actualizar el token", Toast.LENGTH_SHORT).show();
+            }
+        });
+
+
     }
 
     @Override
@@ -132,30 +161,39 @@ public class ListaVendidosPropietarioActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
+    @Override
+    public void onClick(View view) {
+        switch (view.getId()){
+            case R.id.INICIOPROP_agregar_inmueble_BTN:
+                startActivity(new Intent(InicioPropietarioActivity.this, AgregarInmuebleActivity.class));
+                break;
+        }
+    }
+
     private void obtenerLista(){
         databaseFirebase.child("Inmuebles").addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
                 if (snapshot.exists()){
-                    List<InmuebleModel> listaVendidos = new ArrayList<>();
+                    List<InmuebleModel> listaInmuebles = new ArrayList<>();
 
                     for (DataSnapshot snap : snapshot.getChildren()){
-                        String id = snap.getKey();
+                        String id = snap.getKey().toString();
                         String precio = snap.child("precio").getValue().toString();
                         String direccion = (snap.child("direccion").getValue().toString().split(": "))[1];
                         String idProp = snap.child("id_propietario").getValue().toString();
                         String estado = snap.child("estado").getValue().toString();
 
                         if (!idProp.equals(firebaseAuth.getCurrentUser().getUid())) continue;
-                        if (!estado.equals("VENDIDO")) continue;
+                        if (!estado.equals("CREADO")) continue;
 
                         InmuebleModel inmTemp = new InmuebleModel(id, direccion, precio);
 
-                        listaVendidos.add(inmTemp);
+                        listaInmuebles.add(inmTemp);
 
                     }
-                    adaptadorListaVendidos = new InmuebleAdapter(listaVendidos);
-                    listaVendidosRV.setAdapter(adaptadorListaVendidos);
+                    adaptadorListaInmuebles = new InmuebleAdapter(listaInmuebles);
+                    listaInmueblesRV.setAdapter(adaptadorListaInmuebles);
                 }
             }
