+++ app/src/main/java/com/mmw/inmueblelibre/ui/propietario/ListaReservasPropietarioActivity.java
@@ -1,4 +1,4 @@
-package com.mmw.inmueblelibre.UI.cliente;
+package com.mmw.inmueblelibre.ui.propietario;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,6 +14,7 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
@@ -22,18 +23,20 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.mmw.inmueblelibre.UI.global.MainActivity;
+import com.mmw.inmueblelibre.ui.cliente.InicioClienteActivity;
+import com.mmw.inmueblelibre.ui.global.MainActivity;
 import com.mmw.inmueblelibre.R;
-import com.mmw.inmueblelibre.UI.global.ConfiguracionCuentaActivity;
-import com.mmw.inmueblelibre.UI.global.VerDetallesInmuebleActivity;
-import com.mmw.inmueblelibre.UI.propietario.InicioPropietarioActivity;
+import com.mmw.inmueblelibre.ui.global.ConfiguracionCuentaActivity;
+import com.mmw.inmueblelibre.ui.global.VerDetallesInmuebleActivity;
 import com.mmw.inmueblelibre.adapter.InmuebleAdapter;
 import com.mmw.inmueblelibre.model.InmuebleModel;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public class InicioClienteActivity extends AppCompatActivity {
+public class ListaReservasPropietarioActivity extends AppCompatActivity {
 
     Toolbar toolbar;
 
@@ -50,68 +53,83 @@ public class InicioClienteActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_inicio_cliente);
+        setContentView(R.layout.activity_lista_reservas_propietario);
 
-        toolbar = findViewById(R.id.INICIOCLI_toolbar);
+        toolbar = findViewById(R.id.LRP_toolbar);
         setSupportActionBar(toolbar);
-
-        drawerLayout = (DrawerLayout) findViewById(R.id.INICIOCLI_drawer_layout);
-        menuDrawer = (NavigationView) findViewById(R.id.INICIOCLI_menu_drawer);
-
-        listaReservasRV = findViewById(R.id.INICIOCLI_listaReservas);
-        reservasLayoutManager = new LinearLayoutManager(this);
-        listaReservasRV.setLayoutManager(reservasLayoutManager);
-        adaptadorListaReservas = new InmuebleAdapter(new ArrayList<>());
-        listaReservasRV.setAdapter(adaptadorListaReservas);
 
         firebaseAuth = FirebaseAuth.getInstance();
         databaseFirebase = FirebaseDatabase.getInstance().getReference();
 
+        drawerLayout = (DrawerLayout) findViewById(R.id.LRP_drawer_layout);
+        menuDrawer = (NavigationView) findViewById(R.id.LRP_menu_drawer);
+
+        listaReservasRV = findViewById(R.id.LRP_listaInmuebles);
+        reservasLayoutManager = new LinearLayoutManager(this);
+        listaReservasRV.setLayoutManager(reservasLayoutManager);
+        adaptadorListaReservas = new InmuebleAdapter(new ArrayList<>());
+        listaReservasRV.setAdapter(adaptadorListaReservas);
         obtenerLista();
 
-        obtenerInfoUsuario();
-
+        //TODO IMPLEMENTAR ACCIONES DEL MENU DRAWER
         menuDrawer.setNavigationItemSelectedListener(menuItem -> {
 
             drawerLayout.closeDrawer(menuDrawer);
 
-            switch (menuItem.getItemId()) {
-                case R.id.MENUCLI_menu_principal_opc:
+            switch (menuItem.getItemId()){
+                case R.id.MENUPROP_menu_principal_opc:
+                    startActivity(new Intent(ListaReservasPropietarioActivity.this, InicioPropietarioActivity.class));
                     break;
 
-                case R.id.MENUCLI_listar_inmuebles_opc:
-                    startActivity(new Intent(InicioClienteActivity.this, ListaInmueblesClienteActivity.class));
+                case R.id.MENUPROP_listar_reservas_opc:
                     break;
 
-                case R.id.MENUCLI_listar_comprados_opc:
-                    startActivity(new Intent(InicioClienteActivity.this, ListaCompradosClienteActivity.class));
+                case R.id.MENUPROP_listar_vendidos_opc:
+                    startActivity(new Intent(ListaReservasPropietarioActivity.this, ListaVendidosPropietarioActivity.class));
                     break;
 
-                case R.id.MENUCLI_mi_cuenta_opc:
-                    startActivity(new Intent(InicioClienteActivity.this, ConfiguracionCuentaActivity.class));
+                case R.id.MENUPROP_mi_cuenta_opc:
+                    startActivity(new Intent(ListaReservasPropietarioActivity.this, ConfiguracionCuentaActivity.class));
                     break;
 
-                case R.id.MENUCLI_cerrar_sesion_opc:
-                    firebaseAuth.signOut();
-                    startActivity(new Intent(InicioClienteActivity.this, MainActivity.class));
-                    finish();
+                case R.id.MENUPROP_cerrar_sesion_opc:
+                    cerrarSesion();
                     break;
             }
 
             return false;
         });
 
+        obtenerInfoUsuario();
+
         adaptadorListaReservas.setOnItemClickListener(new InmuebleAdapter.OnItemClickListener() {
             @Override
             public void onItemClick(View itemView, int position) {
-                String tipoUsuario = "CLIENTE";
+                String tipoUsuario = "PROPIETARIO";
                 String idInmueble = adaptadorListaReservas.getInmuebleId(position);
 
-                Intent intent = new Intent(InicioClienteActivity.this, VerDetallesInmuebleActivity.class);
+                Intent intent = new Intent(ListaReservasPropietarioActivity.this, VerDetallesInmuebleActivity.class);
                 intent.putExtra("tipo_usuario", tipoUsuario);
                 intent.putExtra("id_inmueble", idInmueble);
                 intent.putExtra("estado_inmueble", "RESERVADO");
                 startActivity(intent);
+            }
+        });
+    }
+
+    private void cerrarSesion(){
+        String id = firebaseAuth.getCurrentUser().getUid();
+
+        Map<String, Object> mapaValores = new HashMap<>();
+        mapaValores.put("token_fcm", "");
+
+        databaseFirebase.child("Usuarios").child(id).updateChildren(mapaValores).addOnCompleteListener(taskDB -> {
+            if (taskDB.isSuccessful()){
+                firebaseAuth.signOut();
+                startActivity(new Intent(ListaReservasPropietarioActivity.this, MainActivity.class));
+                finish();
+            } else {
+                Toast.makeText(getApplicationContext(), "No se pudo actualizar el token", Toast.LENGTH_SHORT).show();
             }
         });
     }
@@ -133,6 +151,40 @@ public class InicioClienteActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
+    private void obtenerLista(){
+        databaseFirebase.child("Inmuebles").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                if (snapshot.exists()){
+                    List<InmuebleModel> listaReservas = new ArrayList<>();
+
+                    for (DataSnapshot snap : snapshot.getChildren()){
+                        String id = snap.getKey();
+                        String precio = snap.child("precio").getValue().toString();
+                        String direccion = (snap.child("direccion").getValue().toString().split(": "))[1];
+                        String idProp = snap.child("id_propietario").getValue().toString();
+                        String estado = snap.child("estado").getValue().toString();
+
+                        if (!idProp.equals(firebaseAuth.getCurrentUser().getUid())) continue;
+                        if (!estado.equals("RESERVADO")) continue;
+
+                        InmuebleModel inmTemp = new InmuebleModel(id, direccion, precio);
+
+                        listaReservas.add(inmTemp);
+
+                    }
+                    adaptadorListaReservas = new InmuebleAdapter(listaReservas);
+                    listaReservasRV.setAdapter(adaptadorListaReservas);
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+
+            }
+        });
+    }
+
     private void obtenerInfoUsuario(){
         String id = firebaseAuth.getCurrentUser().getUid();
 
@@ -144,45 +196,8 @@ public class InicioClienteActivity extends AppCompatActivity {
                     String correo = snapshot.child("email").getValue().toString();
 
                     View drawerHeader = menuDrawer.getHeaderView(0);
-                    ((TextView) drawerHeader.findViewById(R.id.DRAWERCLI_nombre_TV)).setText(nombre);
-                    ((TextView) drawerHeader.findViewById(R.id.DRAWERCLI_correo_TV)).setText(correo);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-
-            }
-        });
-    }
-
-    private void obtenerLista(){
-        databaseFirebase.child("Inmuebles").addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot snapshot) {
-                if (snapshot.exists()){
-                    List<InmuebleModel> listaReservas = new ArrayList<>();
-
-                    for (DataSnapshot snap : snapshot.getChildren()){
-                        String id = snap.getKey().toString();
-
-                        String idActual = firebaseAuth.getCurrentUser().getUid();
-
-                        String idClienteInmueble = snap.child("id_cliente").getValue().toString();
-                        String estado = snap.child("estado").getValue().toString();
-                        String precio = snap.child("precio").getValue().toString();
-                        String direccion = (snap.child("direccion").getValue().toString().split(": "))[1];
-
-                        if (!idActual.equals(idClienteInmueble)) continue;
-                        if (!estado.equals("RESERVADO")) continue;
-
-                        InmuebleModel inmTemp = new InmuebleModel(id, direccion, precio);
-
-                        listaReservas.add(inmTemp);
-                    }
-
-                    adaptadorListaReservas = new InmuebleAdapter(listaReservas);
-                    listaReservasRV.setAdapter(adaptadorListaReservas);
+                    ((TextView) drawerHeader.findViewById(R.id.DRAWERPROP_nombre_TV)).setText(nombre);
+                    ((TextView) drawerHeader.findViewById(R.id.DRAWERPROP_correo_TV)).setText(correo);
                 }
             }
