+++ app/src/main/res/layout/activity_configuracion_cuenta.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.global.ConfiguracionCuentaActivity">
+    tools:context=".ui.global.ConfiguracionCuentaActivity">
 
     <include layout="@layout/toolbar"
         android:id="@+id/CONFIG_toolbar"
