+++ app/src/main/res/layout/activity_inicio_cliente.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".UI.cliente.InicioClienteActivity">
+        tools:context=".ui.cliente.InicioClienteActivity">
 
         <include layout="@layout/toolbar"
             android:id="@+id/INICIOCLI_toolbar"
