+++ app/src/main/res/layout/activity_inicio_propietario.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".UI.propietario.InicioPropietarioActivity">
+        tools:context=".ui.propietario.InicioPropietarioActivity">
 
         <include layout="@layout/toolbar"
             android:id="@+id/INICIOPROP_toolbar"
