+++ app/src/main/res/layout/activity_lista_inmuebles_cliente.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".UI.cliente.ListaInmueblesClienteActivity">
+        tools:context=".ui.cliente.ListaInmueblesClienteActivity">
 
         <include layout="@layout/toolbar"
             android:id="@+id/LIC_toolbar"
