+++ app/src/main/res/layout/activity_lista_reservas_propietario.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".UI.propietario.ListaReservasPropietarioActivity">
+        tools:context=".ui.propietario.ListaReservasPropietarioActivity">
 
         <include layout="@layout/toolbar"
             android:id="@+id/LRP_toolbar"
