+++ app/src/main/res/layout/activity_ver_detalles_inmueble.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UI.global.VerDetallesInmuebleActivity">
+    tools:context=".ui.global.VerDetallesInmuebleActivity">
 
     <include layout="@layout/toolbar"
         android:id="@+id/DETALLES_toolbar"/>
